feat: Update cargo robbery evaluation logic
Refactor the code to include a new function, evaluar_robo, to determine if the cargo is stolen between 2:00 AM and 4:00 AM. If the current time falls within this range, a random number is generated and if it is less than 5, the cargo is considered stolen. The simulation is adjusted to add 60 minutes to the travel time if a robbery occurs.